cart scroller added

+++ src/main/java/org/telegram/galacticMiniatures/bot/model/Listing.java
@@ -26,7 +26,7 @@ public class Listing extends AbstractEntity {
     String description;
 
     @Column(name = "price")
-    Double price;
+    Integer price;
 
     @ManyToOne(fetch = FetchType.LAZY)
     Section section;
